# openapi.yaml
openapi: 3.0.3
info:
  title: FLYFOX AI - NQBA Ecosystem API
  version: "2.0.0"
  description: |
    Complete API for FLYFOX AI's Neuromorphic Quantum Business Architecture (NQBA) ecosystem.
    
    ## Core Components
    
    ### 🎓 FLYFOX Academy
    - Training program enrollment and progress tracking
    - Certification issuance and verification
    - Developer program access and marketplace integration
    
    ### 🔄 Cross-Agent Orchestrator
    - Workflow execution and management
    - Agent registration and capability discovery
    - Cross-agent communication and collaboration
    
    ### ⚡ Quantum Integration Hub (QIH)
    - Quantum job submission and management
    - Optimization problem solving
    - Performance benchmarking and analytics
    
    ### 🏢 Business Units
    - FLYFOX AI (Energy Optimization)
    - Goliath Trade (Financial Intelligence)
    - Sigma Select (Lead Scoring & Sales)
    
    ### 🔐 Authentication & Entitlements
    - JWT-based authentication
    - Role-based access control (RBAC)
    - Feature gating and tier management
    
    ## Technology Stack
    - **Backend**: FastAPI + Python 3.11+
    - **Quantum**: Dynex SDK integration
    - **Database**: PostgreSQL + Redis
    - **Monitoring**: OpenTelemetry + Prometheus + Grafana
    - **Deployment**: Docker + Kubernetes + AWS EKS
    
    ## Getting Started
    1. Obtain API credentials from FLYFOX AI
    2. Authenticate using JWT tokens
    3. Explore available endpoints and capabilities
...
    
...

...
servers:
...
paths:
...
  /orchestrator/workflows/execute:
    post:
      summary: Execute Workflow
      description: Execute a cross-agent workflow
      tags: [Cross-Agent Orchestrator]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowExecutionRequest'
      responses:
        '202':
          description: Workflow execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  workflow_id:
                    type: string
                  workflow_name:
                    type: string
                  status:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /orchestrator/workflows/{workflow_id}/status:
    get:
      summary: Get Workflow Status
      description: Get the current status of a workflow execution
      tags: [Cross-Agent Orchestrator]
      security:
        - bearerAuth: []
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStatusResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Workflow not found
        '500':
          description: Internal server error
  /orchestrator/workflows/{workflow_id}/cancel:
    post:
      summary: Cancel Workflow
      description: Cancel a running workflow
      tags: [Cross-Agent Orchestrator]
      security:
        - bearerAuth: []
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow cancelled successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /orchestrator/agents/register:
    post:
      summary: Register Agent
      description: Register a new agent for orchestration
      tags: [Cross-Agent Orchestrator]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
      responses:
        '201':
          description: Agent registered successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /orchestrator/agents/capabilities:
    get:
      summary: Get Agent Capabilities
      description: Get available agent capabilities, optionally filtered by type
      tags: [Cross-Agent Orchestrator]
      parameters:
        - name: agent_type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AgentType'
      responses:
        '200':
          description: List of agent capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentCapabilityResponse'
        '500':
          description: Internal server error
  /orchestrator/workflows/templates:
    get:
      summary: Get Workflow Templates
      description: Get available workflow templates
      tags: [Cross-Agent Orchestrator]
      responses:
        '200':
          description: List of workflow templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowTemplateResponse'
        '500':
          description: Internal server error
...
  /orchestrator/demo/customer-onboarding:
    get:
      summary: Customer Onboarding Workflow Demo
      description: Demo the customer onboarding workflow
      tags: [Cross-Agent Orchestrator]
      responses:
        '200':
          description: Workflow demo information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDemoResponse'
        '404':
          description: Workflow template not found
        '500':
          description: Internal server error

  # Quantum Integration Hub (QIH) Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  schemas:
...
    AgentType:
      type: string
      enum: [digital_agent, voice_agent, chatbot_agent, quantum_agent, analytics_agent, workflow_agent]
      description: Types of FLYFOX AI agents available for orchestration
    WorkflowStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
      description: Status of cross-agent workflows
    WorkflowExecutionRequest:
      type: object
      required:
        - workflow_name
        - input_data
      properties:
        workflow_name:
          type: string
          description: Name of the workflow template to execute
        input_data:
          type: object
          description: Input data for the workflow
    AgentRegistrationRequest:
      type: object
      required:
        - agent_id
        - agent_type
        - capabilities
        - input_schema
        - output_schema
      properties:
        agent_id:
          type: string
          description: Unique identifier for the agent
        agent_type:
          $ref: '#/components/schemas/AgentType'
        capabilities:
          type: array
          items:
            type: string
          description: List of agent capabilities
        input_schema:
          type: object
          description: JSON schema for agent input
        output_schema:
          type: object
          description: JSON schema for agent output
        execution_timeout:
          type: integer
          default: 300
          description: Execution timeout in seconds
        cost_per_execution:
          type: number
          format: float
          default: 0.0
          description: Cost per execution
    WorkflowStatusResponse:
      type: object
      properties:
        workflow_id:
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        steps_completed:
          type: integer
        total_steps:
          type: integer
        current_step:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    AgentCapabilityResponse:
      type: object
      properties:
        agent_id:
          type: string
        agent_type:
          type: string
        capabilities:
          type: array
          items:
            type: string
        input_schema:
          type: object
        output_schema:
          type: object
        execution_timeout:
          type: integer
        cost_per_execution:
          type: number
          format: float
        availability:
          type: number
          format: float
    WorkflowTemplateResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: object
            properties:
              step_id:
                type: string
              agent_id:
                type: string
              step_type:
                type: string
              input_mapping:
                type: object
              config:
                type: object
...
    WorkflowDemoResponse:
      type: object
      properties:
        message:
          type: string
        workflow_id:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: object
            properties:
              step_id:
                type: string
              agent_type:
                type: string
              description:
                type: string
        demo_input:
          type: object
        estimated_duration:
          type: string

    # Quantum Integration Hub Schemas
...